Add and use a pelias user
For security reasons, it's important that Docker containers run
processes as non-root users where possible. After all, processes in a
Docker container are just regular Linux processes, and we wouldn't run
Pelias processes as the root user outside of Docker.

This PR adds a `pelias` user, which is available for running processes
later.

Because images built with this baseimage may need to install
additional dependencies, we can't run the `USER pelias` Docker command
here, so each of our services/importers needs to do that.

That means this PR should effectively be a no-op until individual
services and importers opt-in to using a non-root user.