Remove nave cache directory
I was looking briefly at the output of [Anchore image
analysis](https://anchore.io/image/dockerhub/pelias%2Fbaseimage:latest#changelog)
for the baseimage, and noticed that nave is storing a cached copy of the
Node.js archive. This is 18MB and is not needed once installed, so it
seems like an easy way to save some space.