Add NPM upgrade step
The Pelias base image currently does not have a step to update NPM.

While using a modern version of Node.js will also ensure a reasonably up
to date version of NPM, there are times when being very up to date is
important.

One of those times is today: NPM just announced a potentially serious
vulnerability in the NPM CLI: https://blog.npmjs.org/post/189618601100/binary-planting-with-the-npm-cli

This change adds an explicit step to upgrade NPM to the latest version
which should help protect against such issues in the future.